Simple script that detects changes in time
This script functions similar to an alarm clock, but much more useless on its own. I intend to use the code here in my own future projects. It is only intended to run on Linux systems, though it could be run on Windows with little modification.

You can set how much in the future that this will go off - in years, months, weeks, and days. It saves the last time it was run in the user's home folder, in a file called ".pyLastRun". It print out "Not quite" if the time in the future hasn't come, and ""!!!!! ALARM ALARM ALARM !!!!! The scheduled thing! Do it!" if it has. These are mainly placeholders. You can change them in lines 140-145 to execute whatever code you need for whichever condition.